chore(expo-font): export `ServerFontResourceDescriptor`
