+++ packages/expo-font/src/ExpoFontLoader.ts
@@ -1,20 +1,6 @@
 import { requireNativeModule } from 'expo-modules-core';
 
-import type { UnloadFontOptions } from './Font.types';
-
-export type ServerFontResourceDescriptor =
-  | {
-      type: 'style';
-      css: string;
-      id: string;
-    }
-  | {
-      type: 'link';
-      as: 'font';
-      crossOrigin?: 'anonymous' | 'use-credentials' | '' | undefined;
-      href: string;
-      rel: 'preload';
-    };
+import type { ServerFontResourceDescriptor, UnloadFontOptions } from './Font.types';
 
 export type ExpoFontLoaderModule = {
   getLoadedFonts: () => string[];
